requirements.txt
characterai
requests
pydub
python-dotenv
websockets
